# api-spec.yaml
openapi: 3.0.3
info:
  title: INFIQAI Labeling Platform API
  description: Privacy-first automated data labeling with Hugging Face integration
  version: 1.0.0
  contact:
    name: INFIQAI Support
    email: support@infiqai.com
servers:
  - url: https://api.infiqai.com/v1
    description: Production server
  - url: https://staging-api.infiqai.com/v1
    description: Staging server
security:
  - BearerAuth: []
paths:
  /datasets:
    get:
      summary: List all datasets
      tags: [Datasets]
      responses:
        '200':
          description: List of datasets
          content:
            application/json:
              schema:
                type: object
                properties:
                  datasets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dataset'
    
    post:
      summary: Create a new dataset
      tags: [Datasets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, taskType, customerId]
              properties:
                name:
                  type: string
                description:
                  type: string
                taskType:
                  type: string
                  enum: [NER, classification, multi-label, contract-tagging]
                customerId:
                  type: string
                confidenceThresholds:
                  $ref: '#/components/schemas/ConfidenceThresholds'
      responses:
        '201':
          description: Dataset created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  dataset:
                    $ref: '#/components/schemas/Dataset'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Dataset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        taskType:
          type: string
          enum: [NER, classification, multi-label, contract-tagging]
        status:
          type: string
          enum: [uploading, processing, ready, completed]
        totalItems:
          type: integer
        labeledItems:
          type: integer
        autoAcceptedItems:
          type: integer
        humanReviewItems:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        customerId:
          type: string
        confidenceThresholds:
          $ref: '#/components/schemas/ConfidenceThresholds'
    ConfidenceThresholds:
      type: object
      properties:
        autoAccept:
          type: number
          minimum: 0
          maximum: 1
        quickReview:
          type: number
          minimum: 0
          maximum: 1
        fullHuman:
          type: number
          minimum: 0
          maximum: 1
...